validator: move create_sidechain_proposal into messages
