+++ src/validator/task.rs
@@ -195,69 +195,6 @@ enum SyncError {
     WriteTxn(#[from] WriteTxnError),
 }
 
-// Returns the serialized sidechain proposal OP_RETURN output.
-fn create_sidechain_proposal(
-    sidechain_number: SidechainNumber,
-    title: String,
-    description: String,
-    hash_1: Vec<u8>,
-    hash_2: Vec<u8>,
-) -> anyhow::Result<Vec<u8>> {
-    const HASH_1_LENGTH: usize = 32;
-    const HASH_2_LENGTH: usize = 20;
-
-    let hash_1: Vec<u8> = if hash_1.is_empty() {
-        vec![0u8; HASH_1_LENGTH]
-    } else if hash_1.len() == HASH_1_LENGTH {
-        hash_1
-    } else {
-        return Err(anyhow::anyhow!(
-            "hash_1 must be empty or {HASH_1_LENGTH} bytes"
-        ));
-    };
-
-    let hash_2: Vec<u8> = if hash_2.is_empty() {
-        vec![0u8; HASH_2_LENGTH]
-    } else if hash_2.len() == HASH_2_LENGTH {
-        hash_2
-    } else {
-        return Err(anyhow::anyhow!(
-            "hash_2 must be empty or {HASH_2_LENGTH} bytes"
-        ));
-    };
-
-    // The only known M1 version.
-    const VERSION: u8 = 0;
-
-    let mut data: Vec<u8> = vec![];
-    data.push(sidechain_number.into());
-    data.push(VERSION);
-    data.push(title.len() as u8);
-    data.extend_from_slice(title.as_bytes());
-    data.extend_from_slice(description.as_bytes());
-    data.extend_from_slice(&hash_1);
-    data.extend_from_slice(&hash_2);
-
-    let Ok(builder) = CoinbaseBuilder::new()
-        .propose_sidechain(sidechain_number.into(), &data)
-        .build()
-    else {
-        return Err(anyhow::anyhow!("Failed to build sidechain proposal"));
-    };
-
-    let tx_out = builder.first().unwrap();
-
-    let mut buffer = Vec::new();
-    match tx_out.consensus_encode(&mut buffer) {
-        Ok(_) => (),
-        Err(e) => {
-            return Err(anyhow::anyhow!("Error encoding tx_out: {}", e));
-        }
-    }
-
-    Ok(buffer)
-}
-
 // See https://github.com/LayerTwo-Labs/bip300_bip301_specifications/blob/master/bip300.md#m1-1
 fn handle_m1_propose_sidechain(
     rwtxn: &mut RwTxn,
@@ -1020,55 +957,4 @@ pub(super) async fn task(
             }
         })
         .await
-}
-#[cfg(test)]
-mod tests {
-
-    use super::*;
-    use bitcoin::{consensus::Decodable, io::Cursor};
-
-    #[test]
-    fn test_roundtrip() {
-        let proposal = create_sidechain_proposal(
-            SidechainNumber::from(13),
-            "title".into(),
-            "description".into(),
-            vec![1u8; 32],
-            vec![2u8; 20],
-        )
-        .expect("Failed to create sidechain proposal");
-
-        let tx_out = TxOut::consensus_decode(&mut Cursor::new(&proposal)).unwrap();
-
-        let (rest, message) = parse_coinbase_script(&tx_out.script_pubkey)
-            .expect("Failed to parse sidechain proposal");
-
-        assert!(rest.is_empty());
-
-        let CoinbaseMessage::M1ProposeSidechain {
-            sidechain_number,
-            data,
-        } = message
-        else {
-            panic!("Failed to parse sidechain proposal");
-        };
-
-        assert_eq!(sidechain_number, 13);
-
-        let proposal = SidechainProposal {
-            sidechain_number: sidechain_number.into(),
-            data,
-            vote_count: 0,
-            proposal_height: 0,
-        };
-
-        let (sidechain_number, deserialized) =
-            (&proposal).try_into().expect("Failed to deserialize");
-
-        assert_eq!(sidechain_number, SidechainNumber(13));
-        assert_eq!(deserialized.description, "description");
-        assert_eq!(deserialized.title, "title");
-        assert_eq!(deserialized.hash_id_1, [1u8; 32]);
-        assert_eq!(deserialized.hash_id_2, [2u8; 20]);
-    }
 }
